+++ examples/cs-oidc-issuer/MyStack.cs
@@ -16,7 +16,7 @@ class MyStack : Pulumi.Stack
             Name = "pulumi_issuer",
             Url = "https://api.pulumi.com/oidc",
             Thumbprints = {
-              "df749a0f34ed673f8b0ec898445910c29c170d01d7d34073bd882235974a8a53"
+              "57d3e89f6b25dde3c174dc558e2b2623306a9d81f88a12e8ae7090a86c12f1da"
             },
             Policies = {
               new AuthPolicyDefinitionArgs
@@ -73,16 +73,6 @@ class MyStack : Pulumi.Stack
               },
               TeamName = "dream-team",
               TokenType = AuthPolicyTokenType.Team
-            },
-            new AuthPolicyDefinitionArgs
-            {
-              Decision = AuthPolicyDecision.Allow,
-              Rules = new Dictionary<string, string> {
-                { "aud", "urn:pulumi:org:"+serviceOrg },
-                { "sub", "repo:organization/repo:*" },
-              },
-              RunnerID = "1234-5678-ABCD-XYZD",
-              TokenType = AuthPolicyTokenType.Runner
             }
           }
         });